Separate file for ColumnFamily and ColumnFamilyDescriptor
