+++ src/db.rs
@@ -728,20 +728,6 @@ impl<'a> Drop for Snapshot<'a> {
         }
     }
 }
-
-impl ColumnFamilyDescriptor {
-    // Create a new column family descriptor with the specified name and options.
-    pub fn new<S>(name: S, options: Options) -> Self
-    where
-        S: Into<String>,
-    {
-        ColumnFamilyDescriptor {
-            name: name.into(),
-            options,
-        }
-    }
-}
-
 impl DB {
     /// Open a database with default options.
     pub fn open_default<P: AsRef<Path>>(path: P) -> Result<DB, Error> {
